Rename FaktaPanel to SideMenuWrapper
Also changes label from "Fakta om" to "Saksopplysninger"
