+++ public/sprak/nb_NO.json
@@ -412,7 +412,7 @@
   "BehandlingHenlagtPanel.Henlagt": "Behandlingen er henlagt",
 
   "TilbakekrevingFakta.FaktaFeilutbetaling": "Feilutbetaling",
-  "FaktaPanel.FaktaOm": "Fakta om",
+  "MainSideMenu.Heading": "Saksopplysninger",
 
   "DashboardResolver.FpLosErNede": "Forsiden har nedetid",
